Move User type to a separate model file and update imports
