+++ models/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+  uid: string;
+  email: string | null;
+  photoURL: string | null;
+  nickname?: string;
+};
